authentication_service/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
requests==2.32.4
pyjwt==2.10.1